feat(explorer-navigation): 中键点击侧边栏在前台打开标签页
- 新增 ExplorerNavigationMiddleClickHook：低级输入观测 + 点击租约隔离 + 原生句柄与 UIA 身份双重关联，中键点击导航窗格文件夹时新标签直接在前台激活

- 新增 MiddleClickForegroundTab 设置项（默认开启），设置界面与托盘菜单可独立开关，不影响合并、复用、双击关闭

- 无障碍接口查询全部在工作线程执行，钩子回调不做阻塞操作；同步处理超过 12ms 即放弃接管并交还原生行为

- 新增 NavigationMiddleClickTests 18 例与 2 项设置持久化测试；修复测试替身 HWND 在 64 位下被截断为 int 的问题

- 版本号更新至 2.0.5

+++ WinTab.Tests/ExplorerTabLifetimeTests.cs
@@ -593,7 +593,7 @@ internal static class ExplorerTabLifetimeTests
                 {
                     "get_HWND" => unreadableHandle
                         ? throw new COMException("The Explorer window is no longer available.")
-                        : (int)parentHandle,
+                        : (long)parentHandle,
                     "get_LocationURL" => unavailableLocation
                         ? throw new COMException("Explorer is temporarily unavailable.")
                         : "file:///C:/WinTab-lifetime",
